requirements.txt
pandas==2.2.0
numpy==1.26.4
spacy==3.7.4
python-dotenv==1.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
